[#19] Rename: pages폴더 정리
