+++ src/pages/price/SendConfirm.tsx
@@ -1,8 +1,8 @@
-import Header from "../components/headers/Header.tsx";
-import Button from "../components/buttons/Button.tsx";
+import Header from "../../components/headers/Header.tsx";
+import Button from "../../components/buttons/Button.tsx";
 import styled from "styled-components";
 import {useNavigate, useParams} from "react-router-dom";
-import mascot from "../assets/home/mascot.svg"
+import mascot from "../../assets/home/mascot.svg"
 import { useEffect, useState } from "react";
 import axios from "axios";
